vertigo-backend/requirements.txt
alchemical==1.0.2
alembic==1.14.0
annotated-types==0.7.0
apifairy==1.4.0
apispec==6.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Faker==33.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
marshmallow-sqlalchemy==1.1.0
mokkari==3.5.0
packaging==24.2
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
ratelimit==2.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
webargs==8.6.0
Werkzeug==3.1.3
